abstracted the js loader a bit more
It's now server-dependent instead of installation-dependent. It takes the base dir (/var/www/html/omeka/base/dir) and strips out the '/var/www/html', which happens to work for both my local dev environment and our server, but it won't on other server setups. It's flimsy but it works for now.